+++ towerhttp/respond_error_test.go
@@ -14,6 +14,22 @@ import (
 	"testing"
 )
 
+type mockNullErrorTransformer struct{}
+
+func (m mockNullErrorTransformer) ErrorBodyTransform(ctx context.Context, err error) any {
+	return nil
+}
+
+type mockErrorEncoder struct{}
+
+func (m mockErrorEncoder) ContentType() string {
+	return ""
+}
+
+func (m mockErrorEncoder) Encode(input any) ([]byte, error) {
+	return nil, errors.New("mock error encoder")
+}
+
 func TestResponder_RespondError(t *testing.T) {
 	type fields struct {
 		encoder          Encoder
@@ -21,12 +37,6 @@ func TestResponder_RespondError(t *testing.T) {
 		errorTransformer ErrorBodyTransformer
 		compressor       Compressor
 		callerDepth      int
-	}
-	type args struct {
-		ctx        context.Context
-		rw         http.ResponseWriter
-		errPayload error
-		opts       []RespondOption
 	}
 	type testRequestGenerator = func(server *httptest.Server) *http.Request
 	towerGen := func(logger tower.Logger) *tower.Tower {
@@ -38,13 +48,13 @@ func TestResponder_RespondError(t *testing.T) {
 		t.SetLogger(logger)
 		return t
 	}
-	//getRequest := func(server *httptest.Server) *http.Request {
-	//	req, err := http.NewRequest(http.MethodGet, server.URL, nil)
-	//	if err != nil {
-	//		t.Fatal(err)
-	//	}
-	//	return req
-	//}
+	getRequest := func(server *httptest.Server) *http.Request {
+		req, err := http.NewRequest(http.MethodGet, server.URL, nil)
+		if err != nil {
+			t.Fatal(err)
+		}
+		return req
+	}
 	postRequest := func(body io.ReadCloser) testRequestGenerator {
 		return func(server *httptest.Server) *http.Request {
 			req, err := http.NewRequest(http.MethodPost, server.URL, body)
@@ -64,7 +74,6 @@ func TestResponder_RespondError(t *testing.T) {
 	tests := []struct {
 		name    string
 		fields  fields
-		args    args
 		server  func(*Responder) *httptest.Server
 		request func(server *httptest.Server) *http.Request
 		test    func(t *testing.T, resp *http.Response, logger *tower.TestingJSONLogger)
@@ -78,7 +87,6 @@ func TestResponder_RespondError(t *testing.T) {
 				compressor:       NoCompression{},
 				callerDepth:      2,
 			},
-			args: args{},
 			server: func(responder *Responder) *httptest.Server {
 				handler := responder.RequestBodyCloner()(http.HandlerFunc(func(writer http.ResponseWriter, request *http.Request) {
 					_, err := io.ReadAll(request.Body)
@@ -159,6 +167,95 @@ func TestResponder_RespondError(t *testing.T) {
 			},
 		},
 		{
+			name: "compressed response",
+			fields: fields{
+				encoder:          NewJSONEncoder(),
+				transformer:      NoopBodyTransform{},
+				errorTransformer: SimpleErrorTransformer{},
+				compressor:       NewGzipCompression(),
+				callerDepth:      2,
+			},
+			server: func(responder *Responder) *httptest.Server {
+				handler := responder.RequestBodyCloner()(http.HandlerFunc(func(writer http.ResponseWriter, request *http.Request) {
+					_, err := io.ReadAll(request.Body)
+					if err != nil {
+						t.Fatalf("failed to read request body: %v", err)
+					}
+					responder.RespondError(writer, request, errors.New(strings.Repeat("test error ", 200)))
+				}))
+				return httptest.NewServer(handler)
+			},
+			request: postRequest(mustJsonBody(map[string]any{"foo": "bar"})),
+			test: func(t *testing.T, resp *http.Response, logger *tower.TestingJSONLogger) {
+				if resp.StatusCode != http.StatusInternalServerError {
+					t.Errorf("expected status code %d, got %d", http.StatusInternalServerError, resp.StatusCode)
+				}
+				if resp.Header.Get("Content-Type") != "application/json" {
+					t.Errorf("expected content type %s, got %s", "application/json", resp.Header.Get("Content-Type"))
+				}
+				body, err := io.ReadAll(resp.Body)
+				if err != nil {
+					t.Fatalf("failed to read response body: %v", err)
+				}
+				if len(body) == 0 {
+					t.Error("expected response body, got empty")
+				}
+				wantBody := `{"error":"<<PRESENCE>>"}`
+				j := jsonassert.New(t)
+				j.Assertf(string(body), wantBody)
+				wantLog := `
+				{
+					"time": "<<PRESENCE>>",
+					"code": 500,
+					"message": "<<PRESENCE>>",
+					"caller": "<<PRESENCE>>",
+					"level": "error",
+					"service": {
+						"name": "responder-test",
+						"environment": "testing",
+						"type": "unit-test"
+					},
+					"context": {
+						"request": {
+							"headers": {
+								"Accept-Encoding": [
+									"gzip"
+								],
+								"User-Agent": [
+									"Go-http-client/1.1"
+								]
+							},
+							"method": "POST",
+							"url": "%s/",
+							"body": {"foo":"bar"}
+						},
+						"response": {
+							"body": {
+								"error": "<<PRESENCE>>"
+							},
+							"headers": {
+								"Content-Encoding": ["gzip"],
+								"Content-Length": [
+									"62"
+								],
+								"Content-Type": [
+									"application/json"
+								]
+							},
+							"status": 500
+						}
+					},
+					"error": {
+						"summary": "<<PRESENCE>>"
+					}
+				}`
+				j.Assertf(logger.String(), wantLog, resp.Request.Host)
+				if !strings.Contains(logger.String(), "towerhttp/respond_error_test.go") {
+					t.Error("expected caller to be in towerhttp/respond_error_test.go")
+				}
+			},
+		},
+		{
 			name: "tower error pattern",
 			fields: fields{
 				encoder:          NewJSONEncoder(),
@@ -167,7 +264,6 @@ func TestResponder_RespondError(t *testing.T) {
 				compressor:       NoCompression{},
 				callerDepth:      2,
 			},
-			args: args{},
 			server: func(responder *Responder) *httptest.Server {
 				handler := responder.RequestBodyCloner()(http.HandlerFunc(func(writer http.ResponseWriter, request *http.Request) {
 					_, err := io.ReadAll(request.Body)
@@ -260,21 +356,468 @@ func TestResponder_RespondError(t *testing.T) {
 				}
 			},
 		},
+		{
+			name: "handled nil error",
+			fields: fields{
+				encoder:          NewJSONEncoder(),
+				transformer:      NoopBodyTransform{},
+				errorTransformer: SimpleErrorTransformer{},
+				compressor:       NoCompression{},
+				callerDepth:      2,
+			},
+			server: func(responder *Responder) *httptest.Server {
+				handler := responder.RequestBodyCloner()(http.HandlerFunc(func(writer http.ResponseWriter, request *http.Request) {
+					_, err := io.ReadAll(request.Body)
+					if err != nil {
+						t.Fatalf("failed to read request body: %v", err)
+					}
+					responder.RespondError(writer, request, nil)
+				}))
+				return httptest.NewServer(handler)
+			},
+			request: postRequest(mustJsonBody(map[string]any{"foo": "bar"})),
+			test: func(t *testing.T, resp *http.Response, logger *tower.TestingJSONLogger) {
+				if resp.StatusCode != http.StatusInternalServerError {
+					t.Errorf("expected status code %d, got %d", http.StatusInternalServerError, resp.StatusCode)
+				}
+				if resp.Header.Get("Content-Type") != "application/json" {
+					t.Errorf("expected content type %s, got %s", "application/json", resp.Header.Get("Content-Type"))
+				}
+				body, err := io.ReadAll(resp.Body)
+				if err != nil {
+					t.Fatalf("failed to read response body: %v", err)
+				}
+				if len(body) == 0 {
+					t.Error("expected response body, got empty")
+				}
+				wantBody := `{"error":"Internal Server Error"}`
+				j := jsonassert.New(t)
+				j.Assertf(string(body), wantBody)
+				wantLog := `
+				{
+					"time": "<<PRESENCE>>",
+					"code": 500,
+					"message": "Internal Server Error",
+					"caller": "<<PRESENCE>>",
+					"level": "error",
+					"service": {
+						"name": "responder-test",
+						"environment": "testing",
+						"type": "unit-test"
+					},
+					"context": {
+						"request": {
+							"headers": {
+								"Accept-Encoding": [
+									"gzip"
+								],
+								"User-Agent": [
+									"Go-http-client/1.1"
+								]
+							},
+							"method": "POST",
+							"url": "%s/",
+							"body": {"foo":"bar"}
+						},
+						"response": {
+							"body": %s,
+							"headers": {
+								"Content-Length": [
+									"<<PRESENCE>>"
+								],
+								"Content-Type": [
+									"application/json"
+								]
+							},
+							"status": 500
+						}
+					},
+					"error": {
+						"summary": "Internal Server Error",
+						"value": "Internal Server Error"
+					}
+				}`
+				j.Assertf(logger.String(), wantLog, resp.Request.Host, wantBody)
+				if !strings.Contains(logger.String(), "towerhttp/respond_error_test.go") {
+					t.Error("expected caller to be in towerhttp/respond_error_test.go")
+				}
+			},
+		},
+		{
+			name: "still received body even when request body is another type of reader",
+			fields: fields{
+				encoder:          NewJSONEncoder(),
+				transformer:      NoopBodyTransform{},
+				errorTransformer: SimpleErrorTransformer{},
+				compressor:       NoCompression{},
+				callerDepth:      2,
+			},
+			server: func(responder *Responder) *httptest.Server {
+				handler := responder.RequestBodyCloner()(http.HandlerFunc(func(writer http.ResponseWriter, request *http.Request) {
+					request.Body = io.NopCloser(request.Body)
+					_, err := io.ReadAll(request.Body)
+					if err != nil {
+						t.Fatalf("failed to read request body: %v", err)
+					}
+					responder.RespondError(writer, request, nil)
+				}))
+				return httptest.NewServer(handler)
+			},
+			request: postRequest(mustJsonBody(map[string]any{"foo": "bar"})),
+			test: func(t *testing.T, resp *http.Response, logger *tower.TestingJSONLogger) {
+				if resp.StatusCode != http.StatusInternalServerError {
+					t.Errorf("expected status code %d, got %d", http.StatusInternalServerError, resp.StatusCode)
+				}
+				if resp.Header.Get("Content-Type") != "application/json" {
+					t.Errorf("expected content type %s, got %s", "application/json", resp.Header.Get("Content-Type"))
+				}
+				body, err := io.ReadAll(resp.Body)
+				if err != nil {
+					t.Fatalf("failed to read response body: %v", err)
+				}
+				if len(body) == 0 {
+					t.Error("expected response body, got empty")
+				}
+				wantBody := `{"error":"Internal Server Error"}`
+				j := jsonassert.New(t)
+				j.Assertf(string(body), wantBody)
+				wantLog := `
+				{
+					"time": "<<PRESENCE>>",
+					"code": 500,
+					"message": "Internal Server Error",
+					"caller": "<<PRESENCE>>",
+					"level": "error",
+					"service": {
+						"name": "responder-test",
+						"environment": "testing",
+						"type": "unit-test"
+					},
+					"context": {
+						"request": {
+							"headers": {
+								"Accept-Encoding": [
+									"gzip"
+								],
+								"User-Agent": [
+									"Go-http-client/1.1"
+								]
+							},
+							"method": "POST",
+							"url": "%s/",
+							"body": {"foo":"bar"}
+						},
+						"response": {
+							"body": %s,
+							"headers": {
+								"Content-Length": [
+									"<<PRESENCE>>"
+								],
+								"Content-Type": [
+									"application/json"
+								]
+							},
+							"status": 500
+						}
+					},
+					"error": {
+						"summary": "Internal Server Error",
+						"value": "Internal Server Error"
+					}
+				}`
+				j.Assertf(logger.String(), wantLog, resp.Request.Host, wantBody)
+				if !strings.Contains(logger.String(), "towerhttp/respond_error_test.go") {
+					t.Error("expected caller to be in towerhttp/respond_error_test.go")
+				}
+			},
+		},
+		{
+			name: "no body is sent when transformer returns nil",
+			fields: fields{
+				encoder:          NewJSONEncoder(),
+				transformer:      NoopBodyTransform{},
+				errorTransformer: mockNullErrorTransformer{},
+				compressor:       NoCompression{},
+				callerDepth:      2,
+			},
+			server: func(responder *Responder) *httptest.Server {
+				handler := responder.RequestBodyCloner()(http.HandlerFunc(func(writer http.ResponseWriter, request *http.Request) {
+					request.Body = io.NopCloser(request.Body)
+					_, err := io.ReadAll(request.Body)
+					if err != nil {
+						t.Fatalf("failed to read request body: %v", err)
+					}
+					responder.RespondError(writer, request, errors.New("foo"))
+				}))
+				return httptest.NewServer(handler)
+			},
+			request: postRequest(mustJsonBody(map[string]any{"foo": "bar"})),
+			test: func(t *testing.T, resp *http.Response, logger *tower.TestingJSONLogger) {
+				if resp.StatusCode != http.StatusInternalServerError {
+					t.Errorf("expected status code %d, got %d", http.StatusInternalServerError, resp.StatusCode)
+				}
+				if resp.Header.Get("Content-Type") != "" {
+					t.Errorf("expected no content type")
+				}
+				body, err := io.ReadAll(resp.Body)
+				if err != nil {
+					t.Fatalf("failed to read response body: %v", err)
+				}
+				if len(body) != 0 {
+					t.Errorf("expected no response body, got %s", string(body))
+				}
+				j := jsonassert.New(t)
+				wantLog := `
+				{
+					"time": "<<PRESENCE>>",
+					"code": 500,
+					"message": "foo",
+					"caller": "<<PRESENCE>>",
+					"level": "error",
+					"service": {
+						"name": "responder-test",
+						"environment": "testing",
+						"type": "unit-test"
+					},
+					"context": {
+						"request": {
+							"headers": {
+								"Accept-Encoding": [
+									"gzip"
+								],
+								"User-Agent": [
+									"Go-http-client/1.1"
+								]
+							},
+							"method": "POST",
+							"url": "%s/",
+							"body": {"foo":"bar"}
+						},
+						"response": {
+							"status": 500
+						}
+					},
+					"error": {
+						"summary": "foo"
+					}
+				}`
+				j.Assertf(logger.String(), wantLog, resp.Request.Host)
+				if !strings.Contains(logger.String(), "towerhttp/respond_error_test.go") {
+					t.Error("expected caller to be in towerhttp/respond_error_test.go")
+				}
+			},
+		},
+		{
+			name: "expected output on error encoding",
+			fields: fields{
+				encoder:          mockErrorEncoder{},
+				transformer:      NoopBodyTransform{},
+				errorTransformer: SimpleErrorTransformer{},
+				compressor:       NoCompression{},
+				callerDepth:      2,
+			},
+			server: func(responder *Responder) *httptest.Server {
+				handler := responder.RequestBodyCloner()(http.HandlerFunc(func(writer http.ResponseWriter, request *http.Request) {
+					request.Body = io.NopCloser(request.Body)
+					_, err := io.ReadAll(request.Body)
+					if err != nil {
+						t.Fatalf("failed to read request body: %v", err)
+					}
+					responder.RespondError(writer, request, errors.New("foo"))
+				}))
+				return httptest.NewServer(handler)
+			},
+			request: getRequest,
+			test: func(t *testing.T, resp *http.Response, logger *tower.TestingJSONLogger) {
+				if resp.StatusCode != http.StatusInternalServerError {
+					t.Errorf("expected status code %d, got %d", http.StatusInternalServerError, resp.StatusCode)
+				}
+				if resp.Header.Get("Content-Type") != "text/plain; charset=utf-8" {
+					t.Errorf("expected content type text/plain; charset=utf-8, got %s", resp.Header.Get("Content-Type"))
+				}
+				body, err := io.ReadAll(resp.Body)
+				if err != nil {
+					t.Fatalf("failed to read response body: %v", err)
+				}
+				if len(body) == 0 {
+					t.Errorf("expected response body")
+				}
+				j := jsonassert.New(t)
+				wantLog := `
+				{
+					"time": "<<PRESENCE>>",
+					"code": 500,
+					"message": "foo",
+					"caller": "<<PRESENCE>>",
+					"level": "error",
+					"service": {
+						"name": "responder-test",
+						"environment": "testing",
+						"type": "unit-test"
+					},
+					"context": {
+						"request": {
+							"headers": {
+								"Accept-Encoding": [
+									"gzip"
+								],
+								"User-Agent": [
+									"Go-http-client/1.1"
+								]
+							},
+							"method": "GET",
+							"url": "%s/"
+						},
+						"response": {
+							"status": 500,
+							"body": "ENCODING ERROR",
+							"headers": {
+								"Content-Type": [
+									"text/plain; charset=utf-8"
+								]
+							}
+						}
+					},
+					"error": {
+						"summary": "foo"
+					}
+				}`
+				j.Assertf(logger.String(), wantLog, resp.Request.Host)
+				if !strings.Contains(logger.String(), "towerhttp/respond_error_test.go") {
+					t.Error("expected caller to be in towerhttp/respond_error_test.go")
+				}
+			},
+		},
+		{
+			name: "still received proper response even compressor fails",
+			fields: fields{
+				encoder: func() Encoder {
+					enc := NewJSONEncoder()
+					enc.SetIndent("")
+					enc.SetPrefix("")
+					enc.SetHtmlEscape(false)
+					return enc
+				}(),
+				transformer:      NoopBodyTransform{},
+				errorTransformer: SimpleErrorTransformer{},
+				compressor:       mockErrorCompressor{},
+				callerDepth:      2,
+			},
+			server: func(responder *Responder) *httptest.Server {
+				handler := responder.RequestBodyCloner()(http.HandlerFunc(func(writer http.ResponseWriter, request *http.Request) {
+					request.Body = io.NopCloser(request.Body)
+					_, err := io.ReadAll(request.Body)
+					if err != nil {
+						t.Fatalf("failed to read request body: %v", err)
+					}
+					responder.RespondError(writer, request, nil)
+				}))
+				return httptest.NewServer(handler)
+			},
+			request: postRequest(mustJsonBody(map[string]any{"foo": "bar"})),
+			test: func(t *testing.T, resp *http.Response, logger *tower.TestingJSONLogger) {
+				if resp.StatusCode != http.StatusInternalServerError {
+					t.Errorf("expected status code %d, got %d", http.StatusInternalServerError, resp.StatusCode)
+				}
+				if resp.Header.Get("Content-Type") != "application/json" {
+					t.Errorf("expected content type %s, got %s", "application/json", resp.Header.Get("Content-Type"))
+				}
+				body, err := io.ReadAll(resp.Body)
+				if err != nil {
+					t.Fatalf("failed to read response body: %v", err)
+				}
+				if len(body) == 0 {
+					t.Error("expected response body, got empty")
+				}
+				wantBody := `{"error":"Internal Server Error"}`
+				j := jsonassert.New(t)
+				j.Assertf(string(body), wantBody)
+				logs := strings.Split(logger.String(), "\n")
+				// 3 split because EOL is \n
+				if len(logs) != 3 {
+					t.Fatalf("expected 2 logs, got %d", len(logs))
+				}
+				wantEntry := `
+				{
+					"time": "<<PRESENCE>>",
+					"code": 500,
+					"message": "compress error",
+					"caller": "<<PRESENCE>>",
+					"level": "warn",
+					"service": {
+						"name": "responder-test",
+						"environment": "testing",
+						"type": "unit-test"
+					},
+					"error": {
+						"summary": "compress error"
+					}
+				}`
+				j.Assertf(logs[0], wantEntry)
+				wantLog := `
+				{
+					"time": "<<PRESENCE>>",
+					"code": 500,
+					"message": "Internal Server Error",
+					"caller": "<<PRESENCE>>",
+					"level": "error",
+					"service": {
+						"name": "responder-test",
+						"environment": "testing",
+						"type": "unit-test"
+					},
+					"context": {
+						"request": {
+							"headers": {
+								"Accept-Encoding": [
+									"gzip"
+								],
+								"User-Agent": [
+									"Go-http-client/1.1"
+								]
+							},
+							"method": "POST",
+							"url": "%s/",
+							"body": {"foo":"bar"}
+						},
+						"response": {
+							"body": %s,
+							"headers": {
+								"Content-Length": [
+									"<<PRESENCE>>"
+								],
+								"Content-Type": [
+									"application/json"
+								]
+							},
+							"status": 500
+						}
+					},
+					"error": {
+						"summary": "Internal Server Error",
+						"value": "Internal Server Error"
+					}
+				}`
+				j.Assertf(logs[1], wantLog, resp.Request.Host, wantBody)
+				if !strings.Contains(logger.String(), "towerhttp/respond_error_test.go") {
+					t.Error("expected caller to be in towerhttp/respond_error_test.go")
+				}
+			},
+		},
 	}
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
 			logger := tower.NewTestingJSONLogger()
 			tow := towerGen(logger)
-			r := Responder{
-				encoder:          tt.fields.encoder,
-				transformer:      tt.fields.transformer,
-				errorTransformer: tt.fields.errorTransformer,
-				tower:            tow,
-				compressor:       tt.fields.compressor,
-				callerDepth:      tt.fields.callerDepth,
-			}
+			r := NewResponder()
+			r.SetEncoder(tt.fields.encoder)
+			r.SetBodyTransformer(tt.fields.transformer)
+			r.SetErrorTransformer(tt.fields.errorTransformer)
+			r.SetTower(tow)
+			r.SetCompressor(tt.fields.compressor)
+			r.SetCallerDepth(tt.fields.callerDepth)
 			r.RegisterHook(NewLoggerHook())
-			server := tt.server(&r)
+			server := tt.server(r)
 			defer server.Close()
 			resp, err := http.DefaultClient.Do(tt.request(server))
 			if err != nil {